Fix onboarding bugs + questionnaire additions
Additions I made to questionnaire are as follows: Elements are now more centered in the middle. There is now a preview that I added for questionnaire. +  Bugs i addressed are as follows: replacing all words anxious with chill. ready to begin is now a card. ready to begin is now a button instead of a hold to start gesture