Implemented all suggestions
Made spacing as close as possible, I couldn't figure out how to control spacing, so I opted for close as possible instead of separated. Got rid of time, but kept date for when it was created. Added swipe functionality and display to user to see whether it was successfully starred or deleted.